# src/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "kuSystem Backend API",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local dev"
    }
  ],
  "components": {
    "schemas": {
      "Client": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1
          },
          "taxId": {
            "type": "string",
            "nullable": true,
            "maxLength": 100
          },
          "phone": {
            "type": "string",
            "nullable": true,
            "maxLength": 100
          },
          "email": {
            "type": "string",
            "nullable": true,
            "maxLength": 200,
            "format": "email"
          },
          "id": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "name",
          "id",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "ClientInputPartial": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1
          },
          "taxId": {
            "type": "string",
            "nullable": true,
            "maxLength": 100
          },
          "phone": {
            "type": "string",
            "nullable": true,
            "maxLength": 100
          },
          "email": {
            "type": "string",
            "nullable": true,
            "maxLength": 200,
            "format": "email"
          }
        }
      },
      "ClientBranch": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "id": {
            "type": "string"
          },
          "clientId": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "name",
          "id",
          "clientId",
          "createdAt",
          "updatedAt"
        ]
      },
      "ClientBranchInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1
          },
          "address": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "name"
        ]
      },
      "ClientBranchInputPartial": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1
          },
          "address": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "Product": {
        "type": "object",
        "properties": {
          "sku": {
            "type": "string",
            "nullable": true,
            "maxLength": 100
          },
          "name": {
            "type": "string",
            "minLength": 1
          },
          "unit": {
            "type": "string",
            "nullable": true,
            "maxLength": 50
          },
          "price": {
            "type": "number",
            "minimum": 0
          },
          "taxRate": {
            "type": "number",
            "nullable": true,
            "minimum": 0,
            "maximum": 1
          },
          "priceIncludesTax": {
            "type": "boolean",
            "default": false
          },
          "id": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "name",
          "price",
          "id",
          "createdAt",
          "updatedAt"
        ]
      },
      "ProductInput": {
        "type": "object",
        "properties": {
          "sku": {
            "type": "string",
            "nullable": true,
            "maxLength": 100
          },
          "name": {
            "type": "string",
            "minLength": 1
          },
          "unit": {
            "type": "string",
            "nullable": true,
            "maxLength": 50
          },
          "price": {
            "type": "number",
            "minimum": 0
          },
          "taxRate": {
            "type": "number",
            "nullable": true,
            "minimum": 0,
            "maximum": 1
          },
          "priceIncludesTax": {
            "type": "boolean",
            "default": false
          }
        },
        "required": [
          "name",
          "price"
        ]
      },
      "ProductInputPartial": {
        "type": "object",
        "properties": {
          "sku": {
            "type": "string",
            "nullable": true,
            "maxLength": 100
          },
          "name": {
            "type": "string",
            "minLength": 1
          },
          "unit": {
            "type": "string",
            "nullable": true,
            "maxLength": 50
          },
          "price": {
            "type": "number",
            "minimum": 0
          },
          "taxRate": {
            "type": "number",
            "nullable": true,
            "minimum": 0,
            "maximum": 1
          },
          "priceIncludesTax": {
            "type": "boolean",
            "default": false
          }
        }
      },
...
    },
...
  },
  "paths": {
...
    "/clients/{id}": {
      "get": {
        "summary": "Get client",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Client",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Client"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update client",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientInputPartial"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Client"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete client",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "ok"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/clients/{clientId}/branches": {
      "get": {
        "summary": "List branches by client",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "clientId",
            "in": "path"
          },
          {
            "schema": {
              "type": "integer"
            },
            "required": false,
            "name": "page",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer"
            },
            "required": false,
            "name": "pageSize",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "List of client branches",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ClientBranch"
                      }
                    },
                    "total": {
                      "type": "number"
                    },
                    "page": {
                      "type": "number"
                    },
                    "pageSize": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "data",
                    "total",
                    "page",
                    "pageSize"
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create client branch",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "clientId",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientBranchInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientBranch"
                }
              }
            }
          }
        }
      }
    },
    "/client-branches/{id}": {
      "get": {
        "summary": "Get branch",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Branch",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientBranch"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update branch",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientBranchInputPartial"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientBranch"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete branch",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "ok"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/products": {
      "get": {
        "summary": "List products",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "search",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer"
            },
            "required": false,
            "name": "page",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer"
            },
            "required": false,
            "name": "pageSize",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "List of products",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Product"
                      }
                    },
                    "total": {
                      "type": "number"
                    },
                    "page": {
                      "type": "number"
                    },
                    "pageSize": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "data",
                    "total",
                    "page",
                    "pageSize"
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create product",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          }
        }
      }
    },
    "/products/{id}": {
      "get": {
        "summary": "Get product",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Product",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update product",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductInputPartial"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete product",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "ok"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  }
}